src/commands: support migrate rebase (#32)
This pull request introduces a new command for rebasing migrations and
integrates it into the CircleCI configuration. The most important
changes include the addition of the new command, updates to the CircleCI
job definitions, and modifications to the CircleCI workflows to include
the new command.

### New Command Addition:
*
[`src/commands/migrate_auto_rebase.yml`](diffhunk://#diff-c8821d3f57f7cc291de693cb91b6c8cdc5354a50996dc034e1f1cee50659f847R1-R27):
Added a new command definition for rebasing migrations with parameters
for `rebase_branch`, `dir`, and `working_directory`.

### CircleCI Job Updates:
*
[`.circleci/test-deploy.yml`](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R191-R211):
Added a new job `command-test-migrate-rebase` to execute the rebase
command.

### CircleCI Workflow Updates:
*
[`.circleci/test-deploy.yml`](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R705-R706):
Updated the `workflows` section to include the new
`command-test-migrate-rebase` job in various workflows.
[[1]](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R705-R706)
[[2]](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R727)
[[3]](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R738)
[[4]](diffhunk://#diff-1d3c0150af90ce1dd4e8ff2ce93160721d21a31790a051a5ab2b6aac49704535R790)